So I have added a function into this code.
It checks for HTML error code 200

And returns webCheck 

It works.

The problem is that I append the result to businessDetail and it is not 
encased in [''] 

See result

When I write the lines I only get the first character of the result.

*******************

[['B L S Computer Solutions'], ['Unit E 6 Basepoint Business Centre'], 
['Luton'], ['LU28DL'], ['http://www.blscomputersolutions.co.uk'], 
'True']

['B L S Computer Solutions', 'Unit E 6 Basepoint Business Centre', 
'Luton', 'LU28DL', 'http://www.blscomputersolutions.co.uk', 'T']